requirements.txt
python-dotenv
requests
nicehash
nordpool
tinytuya
pytz